scripts/images_exporter/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
docker==5.0.2
idna==3.2
PyYAML==5.4.1
requests==2.26.0
urllib3==1.26.7
websocket-client==1.2.1
